thorify: use new provider
