+++ src/index.ts
@@ -1,10 +1,11 @@
 "use strict";
 
 import {extend} from "./extend";
-import { ThorHttpProvider } from "./http-provider";
+// import { ThorHttpProvider } from "./http-provider";
+import { ThorProvider } from "./provider";
 
 const thorify = function(web3Instance: any, host = "http://localhost:8669", timeout= 0) {
-  const provider = new ThorHttpProvider(host, timeout);
+  const provider = new ThorProvider(host, timeout);
   web3Instance.setProvider(provider);
 
   extend(web3Instance);
